+++ README.md
@@ -55,3 +55,10 @@
 - 코드가 좀 더 길어질 순 있으나 이렇게 했을 때 요청 관리에 대한 로직을 따로 분리해서 나중에 재사용 할 수 있다는 장점이 있다.
 
 ## Reducer를 기반으로 커스텀 hook을 만들어서 요청 상태를 훨씬 편하게 관리하는 방법
+
+- 먼저 만들어놓은 reducerEx 함수를 복사한다.
+- src 폴더 하위에 useAsync.js 라는 파일을 만든다.
+- 파일안에 필요한 훅들을 임폴트 해오고 새로운 함수를 만들고 매개변수로 콜백함수와 디펜던시(의존성 배열)을 넣어준다.
+- 콜백함수에는 API를 호출하는 함수를 넣어줄 것이고, 의존성 배열 자리에는 유즈 이펙트를 사용해서 컴포넌트가 로딩 됐을 때,
+- 혹은 어떤 값이 변경 되었을 때, API를 재요청 할 건데, 유즈이펙트에 사용하는 의존성배열 그대로 가져와서 사용해준다.
+-
